requirements.txt
understat
aiohttp==3.9.5
beautifulsoup4
lxml
dash
dash-mantine-components
gunicorn
certifi
aiohttp[speedups]
requests
python-dotenv
